refactor(control-plane): split validate_fix.go by domain
Split validate_fix.go (811 lines) into three focused files:
- validate_fix.go (316 lines): types, validate-fix loop, mutating phase
- validate_fix_enrich.go (185 lines): error enrichment and topic matching
- validate_fix_parse.go (324 lines): Starlark parsing and kwarg helpers

Remove from knownOversizedFiles in architecture tests.
